+++ R/granova-package.R
@@ -1,0 +1,48 @@
+#' @keywords internal
+#' @aliases granova-package granova NULL
+#' @author This package has been written by Robert M. Pruzek and 
+#' James E. Helmreich. Frederic Bertrand is now the maintainer of package and
+#' has been doing several code fixes and improvements to the package.
+#' 
+#' Maintainer: Frederic Bertrand <frederic.bertrand@@lecnam.net>
+#'
+#' @references David C. Hoaglin, Frederick Mosteller and John W. Tukey 
+#' (1991, ISBN:978-0-471-52735-0).
+#' 
+#' Paul R. Rosenbaum (1989)  <doi:10.2307/2684513>.
+#' 
+#' Robert M. Pruzek and James E. Helmreich <https://jse.amstat.org/v17n1/helmreich.html>.
+"_PACKAGE"
+
+#' @importFrom grDevices gray
+#' @importFrom grDevices grey
+#' @importFrom graphics abline
+#' @importFrom graphics axis
+#' @importFrom graphics box
+#' @importFrom graphics identify
+#' @importFrom graphics legend
+#' @importFrom graphics lines
+#' @importFrom graphics mtext
+#' @importFrom graphics par
+#' @importFrom graphics points
+#' @importFrom graphics rug
+#' @importFrom graphics segments
+#' @importFrom graphics symbols
+#' @importFrom graphics title
+#' @importFrom stats aov
+#' @importFrom stats as.formula
+#' @importFrom stats cor
+#' @importFrom stats lm
+#' @importFrom stats pf
+#' @importFrom stats sd
+#' @importFrom stats t.test
+#' @importFrom stats var
+#' @importFrom utils stack
+#' @importFrom car scatter3d 
+#' @importFrom car Identify3d
+#' 
+#' @export granova.1w
+#' @export granova.2w
+#' @export granova.ds
+#' @export granova.contr
+NULL