+++ NEWS.md
@@ -2,4 +2,5 @@
 
 * Maintainer email update
 * Added unit tests
+* Roxygenized namespace
 * Added a `NEWS.md` file to track changes to the package.